basic functionality

+++ src/templates/ToDo_vanilla/blocks/+list/list.js
@@ -9,35 +9,58 @@ const KEY_CODE = {
   ']': 221,
 };
 
-export const createItem = ( list, classList = [ 'list__item', '_level-1' ] ) => {
+export const createItem = ( type, list, classList ) => {
   const item = document.createElement( 'li' );
-  item.classList.add( ...classList );
-    
-  item.addEventListener( 'keydown', createNextItem.bind(null, list) );
-  item.addEventListener( 'keydown', deleteItem );
-  item.addEventListener( 'keydown', shiftItem );
+
+  let checkClassList = classList ? classList : [ `${type}__item`, '_level-1' ];
+
+  item.classList.add( ...checkClassList );
+
+  if ( type !== 'note' ) {
+    item.addEventListener( 'keydown', createNextItem.bind(null, type, list) );
+    item.addEventListener( 'keydown', deleteItem );
+    item.addEventListener( 'keydown', shiftItem );
+  }
+
   item.addEventListener( 'paste', pasteText );
 
   // idk. wanna string
   item.contentEditable = 'true';
 
+  if (type === 'checklist') {
+    const id = `f${(+new Date).toString(16)}`;
+    const checkbox = document.createElement('input');
+    const label = document.createElement('label');
+
+    checkbox.classList.add('checklist__checkbox', 'hidden');
+    checkbox.id = id;
+    checkbox.type = 'checkbox';
+
+    label.classList.add('checklist__label');
+    label.htmlFor = id;
+
+    item.appendChild( checkbox );
+    item.appendChild( label );
+  }
+
+  console.log(item);
   return item;
 };
 
-export const createNextItem = ( list, e ) => {
+export const createNextItem = ( type, list, e ) => {
   if ( !e.shiftKey && e.keyCode === KEY_CODE.ENTER ) {
     e.preventDefault();
 
-    const level = findLevel( e.target );      
-    const item = createItem( list, [ 'list__item', level ] );
-    
+    const level = findLevel( e.target );
+    const item = createItem( type, list, [ `${type}__item`, level ] );
+
     list.insertBefore( item, e.target.nextElementSibling );
     item.focus();
   }
 };
 
 export const deleteItem = e => {
-  if ( 
+  if (
     e.keyCode === KEY_CODE.BACKSPACE
     || e.keyCode === KEY_CODE.DELETE
     || ( e.shiftKey && e.keyCode === KEY_CODE.NUM_DEL )
@@ -62,21 +85,21 @@ export const pasteText = e => {
 export const shiftItem = e => {
   const previous = e.target.previousElementSibling;
 
-  if ( 
-    previous && e.ctrlKey 
+  if (
+    previous && e.ctrlKey
     && ( e.keyCode === KEY_CODE[ '[' ] || e.keyCode === KEY_CODE[ ']' ] )
   ) {
-    
+
     const level = findLevel(e.target);
     const levelPrev = findLevel(previous);
     const numPrev = +levelPrev.slice(-1);
     let num = +level.slice(-1);
-    
+
     if ( e.keyCode === KEY_CODE[ '[' ] ) num = shiftItemLeft( e, num );
     else num = shiftItemRight( e, num, numPrev );
-    
+
     e.target.classList.replace(level, `${level.slice(0, -1)}${num}`);
-    
+
     num - numPrev > 0 ? e.target.classList.add('_reset')
     : e.target.classList.remove('_reset');
   }
@@ -88,14 +111,15 @@ const shiftItemLeft = ( e, num ) =>
 const shiftItemRight = ( e, num, numPrev ) =>
   ( e.keyCode === KEY_CODE[ ']' ] && num < 9 && num - numPrev <= 0 ? ++num : num );
 
-export const createNewList = () => {
+export const createNewList = ( type ) => {
   const formFileds = document.querySelector( '.form__fields' );
   const list = document.createElement( 'ol' );
-  const item = createItem( list );
-  
+  const item = createItem( type, list );
+
   formFileds.appendChild( list );
   list.appendChild( item );
-  list.classList.add('list');
+  list.classList.add( type );
+  list.dataset.key = type;
   item.focus();
 }
