+++ src/templates/ToDo_vanilla/js/index.js
@@ -2,52 +2,36 @@
 
 console.log("i'm running");
 
-import { default as constructor } from '../blocks/list/list';
+import { default as constructor } from '../blocks/+list/list';
 import '../blocks/form/form';
 
 // 1. Кнопки - нотатка, список, чекліст - які показивутимуть відповідну форму
 // 2. 
 
-const initialList = () => {
-  const list = document.querySelector( '.list' );
-  const item = list.querySelector( '.list__item' );
-  
-  item.addEventListener('keydown', constructor.createNextItem.bind(null, list));
-  item.addEventListener('keydown', constructor.deleteItem);
-  item.addEventListener('keydown', constructor.shiftItem);
+
+const initList = ( type ) => {
+
+  const list = document.querySelector( `.${type}` );
+  const item = document.querySelector( `.${type}__item` );
+
+  if ( type !== 'note' ) {
+    item.addEventListener('keydown', constructor.createNextItem.bind(null, type, list));
+    item.addEventListener('keydown', constructor.deleteItem);
+    item.addEventListener('keydown', constructor.shiftItem);
+  }
+
   item.addEventListener('paste', constructor.pasteText);
-};
+}
 
-initialList();
 
-// const input = $('#note');
-// const output = $('#notes');
-// const tNote = $('p',  $('#t-note').content );
-// const submit = $('#submit');
+initList( 'checklist' );
+initList( 'list' );
+initList( 'note' );
 
-// const notes = localStorage.getItem('notes') || '[]';
-// const notesArr = JSON.parse(notes);
+// init checklist checkbox + label
+const checkbox = document.querySelector('.checklist__checkbox');
+const id = `f${(+new Date).toString(16)}`;
+const label = document.querySelector('.checklist__label');
 
-// const createNote = text => {
-//   const note = tNote.cloneNode();
-//   note.textContent = text;
-//   return note;
-// };
-
-// for ( const note of notesArr ) {
-//   output.appendChild( createNote( note ) );
-// }
-
-// submit.addEventListener('click', e => {
-//   e.preventDefault();
-
-//   if (input.value) {
-//     const note = tNote.cloneNode();
-//     note.textContent = input.value;
-//     output.appendChild( note );
-
-//     notesArr.push(input.value);
-//     localStorage.setItem('notes', JSON.stringify(notesArr));
-//     input.value = '';
-//   }
-// });
+checkbox.id = id;
+label.htmlFor = id;
